created storage api and moved all api files to src/apis folder
